Move bluetooth permission requirement to the base flow activity This means the Control Panel will also request the permission
